fixup! separate HoverHighlightCalculator
