+++ src/app/GitUI/UserControls/RevisionGrid/Columns/RevisionGraphColumnProvider.cs
@@ -6,7 +6,6 @@ using GitUI.NBugReports;
 using GitUI.UserControls.RevisionGrid.Graph;
 using GitUI.UserControls.RevisionGrid.Graph.Rendering;
 using GitUIPluginInterfaces;
-using Microsoft;
 
 namespace GitUI.UserControls.RevisionGrid.Columns;
 
@@ -19,20 +18,15 @@ internal sealed class RevisionGraphColumnProvider : ColumnProvider, IDisposable
 
     private int _columnWidth = 0;
 
-    private const int HoverHighlightDebounceMs = 100;
-
-    private readonly record struct HighlightCacheEntry(string GitRefCompleteName, string? GitRefGuid, int RowIndex, VisibleRowRange VisibleRange, int RevisionCount, IReadOnlySet<ObjectId>? HighlightedIds);
-
-    private IReadOnlySet<ObjectId>? _hoverHighlightedIds;
-    private bool _hoverHighlightDirty;
     private VisibleRowRange _cachedVisibleRange;
-    private readonly CancellationTokenSequence _hoverHighlightSequence = new();
+    private readonly HoverHighlightCalculator _hoverHighlight;
 
     public RevisionGraphColumnProvider(RevisionGraph revisionGraph, IGitRevisionSummaryBuilder gitRevisionSummaryBuilder)
         : base("Graph")
     {
         _revisionGraph = revisionGraph;
         _laneInfoProvider = new LaneInfoProvider(new LaneNodeLocator(_revisionGraph), gitRevisionSummaryBuilder);
+        _hoverHighlight = new HoverHighlightCalculator(_revisionGraph, () => _cachedVisibleRange);
 
         Column = new DataGridViewTextBoxColumn
         {
@@ -130,10 +124,9 @@ internal sealed class RevisionGraphColumnProvider : ColumnProvider, IDisposable
             _graphRenderCache.Reset();
         }
 
-        if (_hoverHighlightDirty)
+        if (_hoverHighlight.ConsumeIsDirty())
         {
             // Hover state changed since last render. Reset so all rows re-render with new highlight.
-            _hoverHighlightDirty = false;
             _graphRenderCache.Reset();
         }
 
@@ -216,7 +209,7 @@ internal sealed class RevisionGraphColumnProvider : ColumnProvider, IDisposable
 
             _graphRenderCache.GraphBitmapGraphics.RenderingOrigin = new Point(x, laneRect.Y);
 
-            GraphRenderer.DrawItem(_revisionGraph.Config, _graphRenderCache.GraphBitmapGraphics, rowIndex, rowHeight, _revisionGraph.GetSegmentsForRow, RevisionGraphDrawStyle, _revisionGraph.HeadId, _hoverHighlightedIds);
+            GraphRenderer.DrawItem(_revisionGraph.Config, _graphRenderCache.GraphBitmapGraphics, rowIndex, rowHeight, _revisionGraph.GetSegmentsForRow, RevisionGraphDrawStyle, _revisionGraph.HeadId, _hoverHighlight.HighlightedIds);
         }
     }
 
@@ -229,8 +222,7 @@ internal sealed class RevisionGraphColumnProvider : ColumnProvider, IDisposable
     {
         _graphRenderCache.Reset();
         _graphDisplayCache.Reset();
-        _hoverHighlightedIds = null;
-        _hoverHighlightDirty = true;
+        _hoverHighlight.Clear();
     }
 
     public void HighlightBranch(ObjectId id)
@@ -241,200 +233,15 @@ internal sealed class RevisionGraphColumnProvider : ColumnProvider, IDisposable
     /// <summary>
     ///  Updates the hover highlight to show only the ancestry of the
     ///  <paramref name="gitRef"/> and tracked remote or the tracking local.
-    ///  Debounces by <see cref="HoverHighlightDebounceMs"/> ms before computing,
-    ///  cancelling any prior pending computation when called again.
+    ///  Debounces before computing, cancelling any prior pending computation when called again.
     ///  Set <see langword="null"/> to clear hover highlighting.
     /// </summary>
     /// <param name="gitRef">The ref to highlight, or <see langword="null"/> to clear.</param>
     /// <param name="rowIndex">
     ///  The row index of the hovered ref label, limits the search to the visible range.
     /// </param>
-    public async Task SetHoverHighlightAsync(IGitRef? gitRef, int rowIndex = -1)
-    {
-        CancellationToken cancellationToken = _hoverHighlightSequence.Next();
-        await Task.Delay(HoverHighlightDebounceMs, cancellationToken);
-        ComputeHoverHighlight(gitRef, rowIndex, cancellationToken);
-    }
-
-    private void ComputeHoverHighlight(IGitRef? gitRef, int rowIndex, CancellationToken cancellationToken)
-    {
-        if (gitRef is null || rowIndex < 0)
-        {
-            if (_hoverHighlightedIds is null)
-            {
-                return;
-            }
-
-            _hoverHighlightedIds = null;
-            _hoverHighlightDirty = true;
-            return;
-        }
-
-        if (cancellationToken.IsCancellationRequested)
-        {
-            return;
-        }
-
-        int revisionCount = _revisionGraph.Count;
-        HashSet<ObjectId> ancestorIds = [];
-
-        // Build the set of currently visible ObjectIds below current row (and some parents/children).
-        RevisionGraphRevision? hoveredRevision = _revisionGraph.GetNodeForRow(rowIndex);
-        Validates.NotNull(hoveredRevision);
-        int visibleTo = Math.Max(_cachedVisibleRange.Count - 1, _cachedVisibleRange.FromIndex + _cachedVisibleRange.Count - 1);
-        HashSet<ObjectId> visibleIds = new(capacity: 50 + (2 * _cachedVisibleRange.Count));
-
-        // Add visible ids and their parents, find if a tracked branch is in the set
-        AddIdAndParents(_revisionGraph, rowIndex, visibleIds);
-        bool checkOtherRefs = gitRef.IsRemote || !string.IsNullOrEmpty(gitRef.MergeWith);
-
-        RevisionGraphRevision? belowRev = null;
-        for (int row = rowIndex + 1; row <= visibleTo; row++)
-        {
-            RevisionGraphRevision rev = AddIdAndParents(_revisionGraph, row, visibleIds);
-            if (checkOtherRefs && rev.GitRevision?.Refs.Any(r => IsInBranchGroup(r, gitRef)) is true)
-            {
-                belowRev = rev;
-            }
-        }
-
-        WalkAncestors(hoveredRevision, ancestorIds, visibleIds);
-
-        if (belowRev is not null)
-        {
-            WalkAncestors(belowRev, ancestorIds, visibleIds);
-        }
-        else if (checkOtherRefs)
-        {
-            // Rows above rowIndex are only needed for the upward search
-            int searchFrom = Math.Max(0, _cachedVisibleRange.FromIndex - Math.Max(50, _cachedVisibleRange.Count));
-            for (int row = rowIndex - 1; row >= searchFrom; row--)
-            {
-                RevisionGraphRevision rev = AddIdAndParents(_revisionGraph, row, visibleIds);
-                if (rev.GitRevision?.Refs.Any(r => IsInBranchGroup(r, gitRef)) is true)
-                {
-                    WalkAncestors(rev, ancestorIds, visibleIds);
-                    break;
-                }
-            }
-        }
-
-        IReadOnlySet<ObjectId>? highlightedIds = ancestorIds.Count > 0 ? ancestorIds : null;
-
-        _highlightCache[_highlightCacheNext] = new HighlightCacheEntry(gitRef.CompleteName, gitRef.Guid, rowIndex, _lastVisibleRange, revisionCount, highlightedIds);
-        _highlightCacheNext = (_highlightCacheNext + 1) % HighlightCacheCapacity;
-
-        if (HightlightIdsSameValues(_hoverHighlightedIds, highlightedIds))
-        {
-            return;
-        }
-
-        if (cancellationToken.IsCancellationRequested)
-        {
-            return;
-        }
-
-        _hoverHighlightedIds = highlightedIds;
-        _hoverHighlightDirty = true;
-
-        return;
-
-        static RevisionGraphRevision AddIdAndParents(RevisionGraph revisionGraph, int row, HashSet<ObjectId> visibleIds)
-        {
-            RevisionGraphRevision? rev = revisionGraph.GetNodeForRow(row);
-            Validates.NotNull(rev);
-            visibleIds.Add(rev.Objectid);
-
-            // A segment whose Child is above the visible area still crosses visible rows and must
-            // be highlighted. Include the Parent endpoint of every segment leaving below so that
-            // WalkAncestors can reach them without storing all ancestors.
-            IRevisionGraphRow? graphRow = revisionGraph.GetSegmentsForRow(row);
-            if (graphRow is null)
-            {
-                return rev;
-            }
-
-            foreach (RevisionGraphSegment segment in graphRow.Segments)
-            {
-                visibleIds.Add(segment.Parent.Objectid);
-            }
-
-            return rev;
-        }
-
-        static bool IsInBranchGroup(IGitRef r, IGitRef gitRef)
-            => gitRef.IsTrackingRemote(r)
-            || r.IsTrackingRemote(gitRef)
-
-            // Match NestledRef with grid remote/local
-            // Only gitRef.Guid to not allocate strings
-            || (gitRef.Guid is null && gitRef.CompleteName == r.CompleteName);
-
-        static void WalkAncestors(RevisionGraphRevision revision, HashSet<ObjectId> result, IReadOnlySet<ObjectId> visibleIds)
-        {
-            Stack<RevisionGraphRevision> stack = new();
-            HashSet<ObjectId> visited = [];
-            stack.Push(revision);
-            while (stack.Count > 0)
-            {
-                RevisionGraphRevision current = stack.Pop();
-                if (!visited.Add(current.Objectid))
-                {
-                    continue;
-                }
-
-                // If already in result from a previous WalkAncestors call, all ancestors
-                // of this commit were already processed — prune this entire subtree.
-                if (result.Contains(current.Objectid))
-                {
-                    continue;
-                }
-
-                // Only store revisions that are in the visible set (commit nodes or segment
-                // endpoints crossing the viewport boundary). The visible set is pre-extended
-                // with one segment above and below, so spanning segments are handled.
-                if (visibleIds.Contains(current.Objectid))
-                {
-                    result.Add(current.Objectid);
-                }
-
-                foreach (RevisionGraphRevision parent in current.Parents)
-                {
-                    stack.Push(parent);
-                }
-            }
-        }
-
-        static bool TryApplyFromCache(HighlightCacheEntry[] highlightCache, IGitRef gitRef, int rowIndex, VisibleRowRange lastVisibleRange, int revisionCount, out IReadOnlySet<ObjectId>? hoverHighlightedIds)
-        {
-            for (int i = 0; i < HighlightCacheCapacity; i++)
-            {
-                ref readonly HighlightCacheEntry entry = ref highlightCache[i];
-                if (entry.GitRefCompleteName == gitRef.CompleteName
-                    && entry.GitRefGuid == gitRef.Guid
-                    && entry.RowIndex == rowIndex
-                    && entry.VisibleRange == lastVisibleRange
-                    && entry.RevisionCount == revisionCount)
-                {
-                    hoverHighlightedIds = entry.HighlightedIds;
-                    return true;
-                }
-            }
-
-            hoverHighlightedIds = null;
-            return false;
-        }
-
-        static bool HightlightIdsSameValues(IReadOnlySet<ObjectId>? left, IReadOnlySet<ObjectId>? right)
-        {
-            if (left is null || right is null)
-            {
-                return left is null && right is null;
-            }
-
-            return left.SetEquals(right);
-        }
-    }
+    public Task SetHoverHighlightAsync(IGitRef? gitRef, int rowIndex = -1)
+        => _hoverHighlight.SetAsync(gitRef, rowIndex);
 
     private int CalculateGraphColumnWidth(in VisibleRowRange range)
     {
@@ -457,7 +264,7 @@ internal sealed class RevisionGraphColumnProvider : ColumnProvider, IDisposable
         return false;
     }
 
-    public void Dispose() => _hoverHighlightSequence.Dispose();
+    public void Dispose() => _hoverHighlight.Dispose();
 
     internal TestAccessor GetTestAccessor() => new(this);
 
@@ -472,17 +279,13 @@ internal sealed class RevisionGraphColumnProvider : ColumnProvider, IDisposable
 
         internal GraphCache GraphCache => RevisionGraphColumnProvider._graphRenderCache;
 
-        internal IReadOnlySet<ObjectId>? HoverHighlightedIds => RevisionGraphColumnProvider._hoverHighlightedIds;
-
-        internal bool IsHoverHighlightDirty => RevisionGraphColumnProvider._hoverHighlightDirty;
+        internal HoverHighlightCalculator.TestAccessor HoverHighlight
+            => RevisionGraphColumnProvider._hoverHighlight.GetTestAccessor();
 
         internal void RenderGraphToCache(VisibleRowRange range, int toRowIndex, int rowHeight)
             => RevisionGraphColumnProvider.RenderGraphToCache(range, toRowIndex, rowHeight);
 
         internal void RenderRowToCache(int rowIndex, int rowHeight)
             => RevisionGraphColumnProvider.RenderRowToCache(rowIndex, rowHeight);
-
-        internal void SetHoverHighlight(IGitRef? gitRef, int rowIndex = -1)
-            => RevisionGraphColumnProvider.ComputeHoverHighlight(gitRef, rowIndex, CancellationToken.None);
     }
 }
