+++ tests/app/UnitTests/GitUI.Tests/UserControls/RevisionGrid/Graph/RevisionGraphColumnTests.cs
@@ -145,129 +145,6 @@ public class RevisionGraphColumnTests
         testAccessor.GraphCache.Count.Should().Be(1);
     }
 
-    [Test]
-    public void SetHoverHighlight_should_include_tip_and_ancestors_within_visible_range()
-    {
-        const string tipId = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        const string parentId = "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb";
-        const string rootId = "cccccccccccccccccccccccccccccccccccccccc";
-        const string otherTipId = "dddddddddddddddddddddddddddddddddddddddd";
-
-        RevisionGraph revisionGraph = new();
-        IGitRef main = CreateBranchRef(localName: "main", isHead: true);
-        revisionGraph.Add(CreateRevision(tipId, [parentId], main));
-        revisionGraph.Add(CreateRevision(otherTipId, [rootId], CreateBranchRef(localName: "feature", isHead: true)));
-        revisionGraph.Add(CreateRevision(parentId, [rootId]));
-        revisionGraph.Add(CreateRevision(rootId, []));
-
-        const int rowCount = 4;
-        revisionGraph.CacheTo(rowCount - 1, rowCount - 1);
-
-        RevisionGraphColumnProvider.TestAccessor testAccessor = CreateProvider(revisionGraph);
-        VisibleRowRange range = new(fromIndex: 0, count: rowCount);
-        testAccessor.RenderGraphToCache(range, toRowIndex: rowCount - 1, _rowHeight);
-
-        testAccessor.SetHoverHighlight(main, rowIndex: GetRowForId(revisionGraph, rowCount, tipId));
-
-        testAccessor.HoverHighlightedIds.Should().BeEquivalentTo(
-        [
-            ObjectId.Parse(tipId),
-            ObjectId.Parse(parentId),
-            ObjectId.Parse(rootId),
-        ]);
-    }
-
-    [Test]
-    public void SetHoverHighlight_should_include_one_ancestor_outside_visible_range()
-    {
-        const string tipId = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        const string parentId = "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb";
-        const string rootId = "cccccccccccccccccccccccccccccccccccccccc";
-
-        RevisionGraph revisionGraph = new();
-        IGitRef main = CreateBranchRef(localName: "main", isHead: true);
-        revisionGraph.Add(CreateRevision(tipId, [parentId], main));
-        revisionGraph.Add(CreateRevision(parentId, [rootId]));
-        revisionGraph.Add(CreateRevision(rootId, []));
-
-        const int rowCount = 3;
-        revisionGraph.CacheTo(rowCount - 1, rowCount - 1);
-
-        RevisionGraphColumnProvider.TestAccessor testAccessor = CreateProvider(revisionGraph);
-
-        // Visible range covers only tip and parent; root is scrolled out of view.
-        VisibleRowRange range = new(fromIndex: 0, count: 2);
-        testAccessor.RenderGraphToCache(range, toRowIndex: 1, _rowHeight);
-
-        testAccessor.SetHoverHighlight(main, rowIndex: GetRowForId(revisionGraph, rowCount, tipId));
-
-        testAccessor.HoverHighlightedIds.Should().BeEquivalentTo(
-        [
-            ObjectId.Parse(tipId),
-            ObjectId.Parse(parentId),
-            ObjectId.Parse(rootId),
-        ]);
-    }
-
-    [Test]
-    public void SetHoverHighlight_should_clear_highlighted_ids_when_gitRef_is_null()
-    {
-        const string tipId = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        const string parentId = "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb";
-
-        RevisionGraph revisionGraph = new();
-        IGitRef main = CreateBranchRef(localName: "main", isHead: true);
-        revisionGraph.Add(CreateRevision(tipId, [parentId], main));
-        revisionGraph.Add(CreateRevision(parentId, []));
-
-        const int rowCount = 2;
-        revisionGraph.CacheTo(rowCount - 1, rowCount - 1);
-
-        RevisionGraphColumnProvider.TestAccessor testAccessor = CreateProvider(revisionGraph);
-        VisibleRowRange range = new(fromIndex: 0, count: rowCount);
-        testAccessor.RenderGraphToCache(range, toRowIndex: rowCount - 1, _rowHeight);
-
-        testAccessor.SetHoverHighlight(main, rowIndex: GetRowForId(revisionGraph, rowCount, tipId));
-        testAccessor.HoverHighlightedIds.Should().NotBeNull();
-
-        testAccessor.RenderGraphToCache(range, toRowIndex: rowCount - 1, _rowHeight);
-        testAccessor.IsHoverHighlightDirty.Should().BeFalse();
-
-        testAccessor.SetHoverHighlight(gitRef: null);
-        testAccessor.HoverHighlightedIds.Should().BeNull();
-        testAccessor.IsHoverHighlightDirty.Should().BeTrue();
-    }
-
-    [Test]
-    public void SetHoverHighlight_should_not_mark_cache_dirty_when_hover_selection_is_unchanged()
-    {
-        const string tipId = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-        const string parentId = "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb";
-
-        RevisionGraph revisionGraph = new();
-        IGitRef main = CreateBranchRef(localName: "main", isHead: true);
-        revisionGraph.Add(CreateRevision(tipId, [parentId], main));
-        revisionGraph.Add(CreateRevision(parentId, []));
-
-        const int rowCount = 2;
-        revisionGraph.CacheTo(rowCount - 1, rowCount - 1);
-
-        RevisionGraphColumnProvider.TestAccessor testAccessor = CreateProvider(revisionGraph);
-        VisibleRowRange range = new(fromIndex: 0, count: rowCount);
-        testAccessor.RenderGraphToCache(range, toRowIndex: rowCount - 1, _rowHeight);
-
-        int tipRow = GetRowForId(revisionGraph, rowCount, tipId);
-
-        testAccessor.SetHoverHighlight(main, tipRow);
-        testAccessor.IsHoverHighlightDirty.Should().BeTrue();
-
-        testAccessor.RenderGraphToCache(range, toRowIndex: rowCount - 1, _rowHeight);
-        testAccessor.IsHoverHighlightDirty.Should().BeFalse();
-
-        testAccessor.SetHoverHighlight(main, tipRow);
-        testAccessor.IsHoverHighlightDirty.Should().BeFalse();
-    }
-
     private static void Setup(int rowCount, out RevisionGraphColumnProvider.TestAccessor testAccessor)
         => Setup(rowCount, rowCount, out testAccessor);
 
@@ -297,50 +174,5 @@ public class RevisionGraphColumnTests
         testAccessor.GraphCache.Reset();
         testAccessor.GraphCache.Capacity.Should().Be(expectedCapacity);
         testAccessor.GraphCache.Count.Should().Be(0);
-    }
-
-    private static int GetRowForId(RevisionGraph revisionGraph, int rowCount, string id)
-    {
-        ObjectId objectId = ObjectId.Parse(id);
-        for (int row = 0; row < rowCount; row++)
-        {
-            if (revisionGraph.GetNodeForRow(row)?.Objectid == objectId)
-            {
-                return row;
-            }
-        }
-
-        throw new InvalidOperationException($"No row found for revision {id}.");
-    }
-
-    private static IGitRef CreateBranchRef(string localName, bool isHead, bool isRemote = false)
-    {
-        IGitRef gitRef = Substitute.For<IGitRef>();
-        gitRef.IsHead.Returns(isHead);
-        gitRef.IsRemote.Returns(isRemote);
-        gitRef.LocalName.Returns(localName);
-        return gitRef;
-    }
-
-    private static RevisionGraphColumnProvider.TestAccessor CreateProvider(RevisionGraph revisionGraph)
-    {
-        IGitRevisionSummaryBuilder gitRevisionSummaryBuilder = Substitute.For<IGitRevisionSummaryBuilder>();
-        RevisionGraphColumnProvider revisionGraphColumnProvider = new(revisionGraph, gitRevisionSummaryBuilder);
-        return revisionGraphColumnProvider.GetTestAccessor();
-    }
-
-    private static GitRevision CreateRevision(string id, IReadOnlyList<string> parentIds, params IGitRef[] refs)
-    {
-        List<ObjectId> parsedParentIds = [];
-        foreach (string parentId in parentIds)
-        {
-            parsedParentIds.Add(ObjectId.Parse(parentId));
-        }
-
-        return new GitRevision(ObjectId.Parse(id))
-        {
-            ParentIds = parsedParentIds,
-            Refs = refs,
-        };
     }
 }
